refactor: split train agents entities into own files, added scala docs
